Add LibraryComparableSortModifier
- much like search moving this up and out allows clients to be able to attach it where it will work for them.
