+++ Sources/Site/Music/UI/LibraryComparableList.swift
@@ -51,28 +51,6 @@ where T: LibraryComparable, T: Identifiable, T: Hashable, T.ID == String {
       }
     }
     .listStyle(.plain)
-    .toolbar {
-      ToolbarItem(placement: .primaryAction) {
-        let sortText = Text(
-          "Sort", bundle: .module,
-          comment: "Shown to change the sort order of the LibraryComparableList.")
-        Menu {
-          Picker(selection: $algorithm) {
-            ForEach(LibrarySectionAlgorithm.allCases, id: \.self) { category in
-              Text(category.localizedString).tag(category)
-            }
-          } label: {
-            sortText
-          }
-        } label: {
-          Label {
-            sortText
-          } icon: {
-            Image(systemName: "line.3.horizontal.decrease.circle")
-          }
-        }
-      }
-    }
   }
 }
